+++ CorgEng.Networking/Networking/Client/NetworkingClient.cs
@@ -40,7 +40,7 @@ namespace CorgEng.Networking.Networking.Client
     /// to iron out some of the oversights.
     /// </summary>
     [Dependency]
-    internal class NetworkingClient : INetworkingClient
+    internal class NetworkingClient : NetworkCommunicator, INetworkingClient
     {
 
         [UsingDependency]
@@ -61,79 +61,9 @@ namespace CorgEng.Networking.Networking.Client
         [UsingDependency]
         private static IPrototypeManager PrototypeManager;
 
-        private IPacketQueue PacketQueue;
-
         public event ConnectionSuccess OnConnectionSuccess;
 
         public event ConnectionFailed OnConnectionFailed;
-
-        public event NetworkMessageRecieved NetworkMessageReceived;
-
-        /// <summary>
-        /// The client that we are using to communicate
-        /// </summary>
-        private UdpClient udpClient;
-
-        /// <summary>
-        /// The address we are connected to, if any
-        /// </summary>
-        private IPAddress address;
-
-        /// <summary>
-        /// The port we are connected to
-        /// </summary>
-        private int port;
-
-        /// <summary>
-        /// Are we connected to a server?
-        /// </summary>
-        private bool connected;
-
-        /// <summary>
-        /// Are we connecting to a server?
-        /// </summary>
-        private bool connecting;
-
-        /// <summary>
-        /// Are we running, or should we shutdown
-        /// </summary>
-        private volatile bool running = false;
-
-        /// <summary>
-        /// The trigger that gets activated when shutdown is triggered
-        /// </summary>
-        private readonly AutoResetEvent shutdownThreadTrigger = new AutoResetEvent(false);
-
-        /// <summary>
-        /// The countdown event that allows the shutdown method to wait until threads are shutdown.
-        /// </summary>
-        private readonly CountdownEvent shutdownCountdown = new CountdownEvent(2);
-
-        /// <summary>
-        /// Have we ever been started?
-        /// </summary>
-        private bool started = false;
-
-        /// <summary>
-        /// If this is true when shutdownThreadTrigger is raised, task will immediately end.
-        /// </summary>
-        private bool cancelToken = false;
-
-        public int TickRate { get; set; } = 32;
-
-        /// <summary>
-        /// The queue of packets that we have sent, so that we can check after some time
-        /// that the packet actually arrived like intended.
-        /// A priority queue sorted in order of when the packet will timeout.
-        /// </summary>
-        public PriorityQueue<IQueuedPacket, double> packetConfirmationQueue = new PriorityQueue<IQueuedPacket, double>();
-
-        [UsingDependency]
-        private static IQueuedPacketFactory QueuedPacketFactory = null!;
-
-#if DEBUG
-        private static Random random = new Random();
-#endif
 
         /// <summary>
         /// Attempt connection to a network address
@@ -267,196 +197,9 @@ namespace CorgEng.Networking.Networking.Client
         }
 
         /// <summary>
-        /// The sender thread.
-        /// Runs when it needs to, transmits data to the server
-        /// with a set tick rate.
-        /// </summary>
-        private void NetworkSenderThread(UdpClient client)
-        {
-            Logger?.WriteLine($"Client sender for {address} thread successfull started.", LogType.DEBUG);
-            Stopwatch stopwatch = new Stopwatch();
-            double inverseTickrate = 1000.0 / TickRate;
-            while (running && (client.Client?.Connected ?? false))
-            {
-                try
-                {
-                    //Create a stopwatch to get the current time
-                    stopwatch.Restart();
-                    //Transmit packets
-                    while (PacketQueue.AcquireLockIfHasMessages())
-                    {
-                        try
-                        {
-                            //Dequeue the packet from the queue
-                            IQueuedPacket queuedPacket = PacketQueue.DequeuePacket();
-                            //Transmit the packet to the server
-                            byte[] data = queuedPacket.Data;
-                            //Asynchronously send the data
-                            //We send all the data straight to the server.
-                            //The client cannot communicate with other clients.
-                            udpClient.SendAsync(data, queuedPacket.TopPointer);
-                            //Register the packet so that after a set amount of time
-                            //we can resend it if it didn't arrive.
-                            OnPacketSent(queuedPacket);
-                        }
-                        finally
-                        {
-                            PacketQueue.ReleaseLock();
-                        }
-                        //Logger.WriteLine($"Sent packets to the server", LogType.TEMP);
-                    }
-                    // Check the packet confirmation queue, to see if we need to resent things
-                    while (packetConfirmationQueue.TryPeek(out var queuedPacket, out double resendTime) && resendTime > CorgEngMain.Time)
-                    {
-                        // Remove the packet
-                        packetConfirmationQueue.Dequeue();
-                        // Resend the packet
-
-                    }
-                    //Wait for variable time to maintain the tick rate
-                    stopwatch.Stop();
-                    double elapsedMilliseconds = stopwatch.ElapsedMilliseconds;
-                    double waitTime = inverseTickrate - elapsedMilliseconds;
-                    if (waitTime > 0)
-                    {
-                        //Sleep the thread
-                        shutdownThreadTrigger.WaitOne((int)waitTime);
-                    }
-                }
-                catch (Exception e)
-                {
-                    Logger?.WriteLine(e, LogType.ERROR);
-                }
-            }
-            Logger?.WriteLine($"Client sender thread terminated", LogType.WARNING);
-            shutdownCountdown.Signal();
-        }
-
-        /// <summary>
-        /// Called when a packet is sent, adds it to the queue of things waiting to be verified that they were sent
-        /// successfully.
-        /// </summary>
-        /// <param name="sentPacket"></param>
-        private void OnPacketSent(IQueuedPacket sentPacket)
-        {
-            sentPacket.SentAt = CorgEngMain.Time;
-            //TODO: Implement ping (For now allow 100ms before a packet is resent. Should be probably around 20-50% the return time ping.)
-            packetConfirmationQueue.Enqueue(sentPacket, sentPacket.SentAt + 0.1);
-            //packetConfirmationQueue.Add(sentPacket.PacketIdentifier, sentPacket);
-        }
-
-        private ConcurrentQueue<(IPEndPoint, byte[])> packetQueue = new ConcurrentQueue<(IPEndPoint, byte[])>();
-
-        /// <summary>
-        /// The networking thread
-        /// </summary>
-        private void NetworkListenerThread(UdpClient client)
-        {
-            Logger?.WriteLine($"Client listener for {address} thread successfull started.", LogType.DEBUG);
-            //Continue always
-            while (running && (client.Client?.Connected ?? false))
-            {
-                try
-                {
-                    //Wait until we are woken up
-                    IPEndPoint remoteEndPoint = new IPEndPoint(IPAddress.Any, port);
-                    byte[] incomingData = udpClient.Receive(ref remoteEndPoint);
-                    //Handle incomming data
-                    packetQueue.Enqueue((remoteEndPoint, incomingData));
-                }
-                catch (Exception e)
-                {
-                    Logger?.WriteLine(e, LogType.ERROR);
-                }
-            }
-            //Disconnected.
-            Logger?.WriteLine("Disconnected from remote server.", LogType.WARNING);
-            shutdownCountdown.Signal();
-        }
-
-        private void ProcessQueueThread()
-        {
-            Logger?.WriteLine($"Packet queue processor thread started.", LogType.MESSAGE);
-            while (running)
-            {
-                //Nothing to process
-                if (packetQueue.Count == 0)
-                {
-                    Thread.Yield();
-                    continue;
-                }
-                //Stuff to do
-                try
-                {
-                    //Recieve messages
-                    (IPEndPoint, byte[]) packet;
-                    if (packetQueue.TryDequeue(out packet))
-                    {
-                        ProcessPacket(packet.Item1, packet.Item2);
-                    }
-                }
-                catch (Exception e)
-                {
-                    Logger?.WriteLine($"Critical exception on processing thread: {e}", LogType.ERROR);
-                }
-            }
-            //Log shutdown
-            Logger?.WriteLine($"Packet queue processor thread stopped.", LogType.MESSAGE);
-        }
-
-        private void ProcessPacket(IPEndPoint sender, byte[] data)
-        {
-            try
-            {
-#if DEBUG
-                if (random.NextDouble() * 100 < NetworkConfig.PacketDropProbability)
-                    return;
-#endif
-                //Ignore messages from people we weren't connecting to.
-                //All communications must go through the server.
-                //This is for security reasons, so a hacked client can't tell other players
-                //invalid information.
-                if (!sender.Address.Equals(address))
-                {
-                    return;
-                }
-                Logger.WriteMetric("packet_size", data.Length.ToString());
-                bool needsAcknowledgement = true;
-                //Convert the packet into the individual messages
-                int messagePointer = 8;
-                while (messagePointer < data.Length)
-                {
-                    int originalPoint = messagePointer;
-                    //Read the integer (First 4 bytes is the size of the message)
-                    int packetSize = BitConverter.ToInt16(data, originalPoint);
-                    //Move the message pointer along
-                    messagePointer += packetSize + 0x06;
-                    //Read the packet header
-                    PacketHeaders packetHeader = (PacketHeaders)BitConverter.ToInt32(data, originalPoint + 0x02);
-                    // Send an acknowledgement packet if we need to
-                    if (needsAcknowledgement && packetHeader != PacketHeaders.ACKNOWLEDGE_PACKET)
-                    {
-                        // Send the acknowledgement request
-                        // Read the packet ID
-                        double packetId = BitConverter.ToInt64(data, 0);
-                        // Tell the server that we recieved the packet
-                        QueueMessage(NetworkMessageFactory.CreateMessage(PacketHeaders.ACKNOWLEDGE_PACKET, BitConverter.GetBytes(packetId)));
-                        needsAcknowledgement = false;
-                    }
-                    //Get the data and pass it on
-                    HandleMessage(sender, packetHeader, data, originalPoint + 0x06, packetSize);
-                }
-            }
-            catch (Exception e)
-            {
-                Logger?.WriteLine(e, LogType.ERROR);
-            }
-        }
-
-        /// <summary>
         /// Handle an incoming message
         /// </summary>
-        private void HandleMessage(IPEndPoint sender, PacketHeaders header, byte[] data, int start, int length)
+        protected override void HandleMessage(IPEndPoint sender, PacketHeaders header, byte[] data, int start, int length)
         {
             Logger.WriteMetric("message_size", length.ToString());
             Logger.WriteMetric("message_header", header.ToString());
@@ -498,9 +241,6 @@ namespace CorgEng.Networking.Networking.Client
             }
             else
             {
-                //Logger?.WriteLine($"Receieved client message: {header} from {sender.Address}", LogType.LOG);
-                //Handle connected packets
-                NetworkMessageReceived?.Invoke(header, data, start, length);
                 //Handle the event
                 switch (header)
                 {
@@ -614,32 +354,12 @@ namespace CorgEng.Networking.Networking.Client
             }
         }
 
-        public void Cleanup()
+        public override void Cleanup()
         {
-            cancelToken = true;
-            Logger?.WriteLine("Client cleanup called", LogType.LOG);
-            running = false;
-            shutdownThreadTrigger.Set();
-            udpClient?.Close();
-            udpClient?.Dispose();
-            udpClient = null;
-            connected = false;
-            connecting = false;
-            NetworkMessageReceived = null;
             //OnConnectionFailed = null;
             OnConnectionSuccess = null;
-            NetworkConfig.ProcessClientSystems = false;
-            if (!NetworkConfig.ProcessServerSystems)
-                NetworkConfig.NetworkingActive = false;
-            Logger?.WriteLine("Waiting for client cleanup completion...", LogType.LOG);
-            //Wait for the threads to be closed
-            if (started)
-                shutdownCountdown.Wait();
-            //Reset
-            shutdownThreadTrigger.Reset();
-            started = false;
-            cancelToken = false;
-            Logger?.WriteLine("Client cleanup completed!", LogType.LOG);
+            base.Cleanup();
         }
+
     }
 }
